CODE/DL MODELS (copy)/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-image>=0.21.0
tqdm>=4.65.0
ultralytics>=8.0.0
albumentations>=1.3.0
tensorboard>=2.13.0
segmentation-models-pytorch>=0.3.3